init: move cpu initialization out of init
There is no need for the cpu and idle thread initialization to be
crammed into init.c. Move out to new/existing files for readability.
